[FIX] Fixes categorical contrasts in python  (#250)
* [ENH] Version 0.3.0 (#249)

- Fixed a bug where the Schaefer parcellation would use the same numbers for the left/right hemisphere.
- Changed the SLM.P peak/clus output to DataFrames/tables. 
- Added peak/clus to the tutorials. 
- Removed some clutter.

* Tutorial update,categorical variables not unsigned

+++ brainstat/stats/terms.py
@@ -84,7 +84,7 @@ def to_df(
             )
         df.columns = names
 
-    return pd.get_dummies(df)
+    return pd.get_dummies(df, dtype=np.int8)
 
 
 def get_index(df: pd.DataFrame) -> Optional[int]:
